requirements.txt
keras==2.9.0
keras_vggface==0.6
matplotlib==3.7.1
mtcnn==0.1.1
numpy==1.24.2
opencv_python==4.7.0.72
pandas==2.0.0
Pillow==9.5.0
tensorflow==2.9.0
tqdm==4.65.0
